Your gateway gave a 502, adding an extra gateway
Local ipfs was down, your gateway gave a 502 and both gateway.ipfs.io and ipfw.pics had problems with cross-origin policies.

I just stated a new gateway just to see how it works and to try it with a Kodi plugin I'm developing. Can't say it will be up forever, but it might help to add the gw anyway.